More flexible way to create app needed for tesating. Requires change in dir structure
