+++ checker/templates/check_of_all_indentified_entities.html
